feat: Integrate AST nesting depth analyzer into metrics
- Moved `NestingDepthVisitor` and `calculate_max_depth` from `src/agent_scorecard/analyzer.py` to `src/agent_scorecard/metrics.py` to resolve potential circular imports and enable function-level analysis.
- Updated `src/agent_scorecard/types.py` to include `nesting_depth` in `FunctionMetric`.
- Updated `get_function_stats` in `src/agent_scorecard/metrics.py` to calculate and return `nesting_depth` for each function.
- Updated `src/agent_scorecard/analyzer.py` to re-export the moved symbols for backward compatibility.
- Added comprehensive tests in `tests/test_analyzer_new.py` verifying the integration and correctness of nesting depth calculation.
